# src/product-service/docs/swagger.yaml
definitions:
  common.AppError:
    properties:
      key:
        type: string
      log:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      filter: {}
      message: {}
      paging: {}
    type: object
  models.AddCategory:
    properties:
      category_name:
        type: string
      description:
        type: string
    type: object
...
  models.Category:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      description:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/categories:
    get:
      consumes:
      - application/json
      description: Retrieve all products, with optional filtering and paging
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: '-'
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Retrieve a list of categories
      tags:
      - categories
  /api/category:
    post:
      consumes:
      - application/json
      description: Create a new category with the provided parameters
      parameters:
      - description: Category to create
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.AddCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Create a new category
      tags:
      - categories
  /api/category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Deleted a category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get a category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Get a category
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category with the provided parameters
      parameters:
      - description: Category to update
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Update an existing category
      tags:
      - categories
...
swagger: "2.0"